tests/wpt/web-platform-tests/css/css-masking/parsing/mask-type-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask-type with invalid values</title>
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-mask-type">
<meta name="assert" content="mask-type supports only the grammar 'luminance | alpha'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("mask-type", "auto");
test_invalid_value("mask-type", "luminance alpha");
test_invalid_value("mask-type", "alpha, luminance");
</script>
</body>
</html>
